+++ game/ato/flightplans/bai.py
@@ -1,0 +1,36 @@
+from __future__ import annotations
+
+from typing import Type
+
+from game.theater.theatergroundobject import TheaterGroundObject
+from .formationattack import FormationAttackBuilder, FormationAttackFlightPlan
+from .invalidobjectivelocation import InvalidObjectiveLocation
+from .waypointbuilder import StrikeTarget
+from ..flightwaypointtype import FlightWaypointType
+
+
+class BaiFlightPlan(FormationAttackFlightPlan):
+    @staticmethod
+    def builder_type() -> Type[Builder]:
+        return Builder
+
+
+class Builder(FormationAttackBuilder[BaiFlightPlan]):
+    def build(self) -> FormationAttackFlightPlan:
+        location = self.package.target
+
+        from game.transfers import Convoy
+
+        targets: list[StrikeTarget] = []
+        if isinstance(location, TheaterGroundObject):
+            for group in location.groups:
+                if group.units:
+                    targets.append(
+                        StrikeTarget(f"{group.group_name} at {location.name}", group)
+                    )
+        elif isinstance(location, Convoy):
+            targets.append(StrikeTarget(location.name, location))
+        else:
+            raise InvalidObjectiveLocation(self.flight.flight_type, location)
+
+        return self._build(BaiFlightPlan, FlightWaypointType.INGRESS_BAI, targets)
