+++ game/ato/flightplans/ocarunway.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+import logging
+from typing import Type
+
+from game.theater import Airfield
+from .formationattack import FormationAttackBuilder, FormationAttackFlightPlan
+from .invalidobjectivelocation import InvalidObjectiveLocation
+from ..flightwaypointtype import FlightWaypointType
+
+
+class OcaRunwayFlightPlan(FormationAttackFlightPlan):
+    @staticmethod
+    def builder_type() -> Type[Builder]:
+        return Builder
+
+
+class Builder(FormationAttackBuilder[OcaRunwayFlightPlan]):
+    def build(self) -> FormationAttackFlightPlan:
+        location = self.package.target
+
+        if not isinstance(location, Airfield):
+            logging.exception(
+                f"Invalid Objective Location for OCA/Runway flight "
+                f"{self.flight=} at {location=}."
+            )
+            raise InvalidObjectiveLocation(self.flight.flight_type, location)
+
+        return self._build(OcaRunwayFlightPlan, FlightWaypointType.INGRESS_OCA_RUNWAY)
