+++ game/ato/flightplans/planningerror.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+
+class PlanningError(RuntimeError):
+    """Raised when the flight planner was unable to create a flight plan."""
